test(opal): split notebook review contracts
