+++ .secrets.baseline
@@ -148,14 +148,14 @@
         "filename": "src/dnadesign/opal/tests/notebooks/test_notebook_template.py",
         "hashed_secret": "71c89bdc763ce580638322b5a3119f484698bf4b",
         "is_verified": false,
-        "line_number": 2943
+        "line_number": 2732
       },
       {
         "type": "Base64 High Entropy String",
         "filename": "src/dnadesign/opal/tests/notebooks/test_notebook_template.py",
         "hashed_secret": "4137adb6e7d947f707c7ff7c44ca11d3586e3eb4",
         "is_verified": false,
-        "line_number": 2944
+        "line_number": 2733
       }
     ],
     "src/dnadesign/usr/assets/demo_material/demo_attachment_one.csv": [
@@ -186,5 +186,5 @@
       }
     ]
   },
-  "generated_at": "2026-08-16T02:03:16Z"
+  "generated_at": "2026-08-16T13:23:29Z"
 }
